move sst to domain for split of engine store
