+++ engine_store_ffi/src/ffi/engine_store_helper_impls.rs
@@ -5,6 +5,7 @@ use kvproto::{kvrpcpb, metapb, raft_cmdpb};
 
 use super::{
     basic_ffi_impls::*,
+    domain_impls::*,
     interfaces_ffi,
     interfaces_ffi::{
         BaseBuffView, ColumnFamilyType, CppStrWithView, EngineStoreApplyRes,
@@ -13,7 +14,6 @@ use super::{
         RawCppStringPtr, RawVoidPtr, SpecialCppPtrType, StoreStats, RAFT_STORE_PROXY_MAGIC_NUMBER,
         RAFT_STORE_PROXY_VERSION,
     },
-    sst_reader_impls::*,
     UnwrapExternCFunc, WriteCmds,
 };
